+++ doc/changelog.rst
@@ -6,6 +6,8 @@ Version 0.5
 
 Changes:
 
+* Replace mean with median
+* Add :meth:`perf.Benchmark.median` method
 * ``Benchmark.get_metadata()`` method removed: use directly the
   :attr:`perf.Benchmark.metadata` attribute